refactor: Store도메인의 Repository 정리
