+++ api/src/main/java/com/albert/commerce/store/infra/presentation/StoreDaoImpl.java
@@ -1,19 +1,26 @@
-package com.albert.commerce.store.infra;
+package com.albert.commerce.store.infra.presentation;
 
 import com.albert.commerce.store.command.domain.QStore;
 import com.albert.commerce.store.command.domain.Store;
 import com.albert.commerce.store.command.domain.StoreId;
+import com.albert.commerce.store.infra.presentation.imports.StoreJpaRepository;
+import com.albert.commerce.store.query.StoreDao;
 import com.albert.commerce.store.ui.StoreNotFoundException;
 import com.albert.commerce.user.command.domain.QUser;
 import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
 
 @RequiredArgsConstructor
-public class StoreDaoCustomImpl implements StoreDaoCustom {
+@Repository
+public class StoreDaoImpl implements StoreDao {
 
     private final JPAQueryFactory jpaQueryFactory;
+    private final StoreJpaRepository storeJpaRepository;
 
+    @Override
     public Store findStoreByUserEmail(String email) {
         QUser user = QUser.user;
         QStore store = QStore.store;
@@ -32,6 +39,7 @@ public class StoreDaoCustomImpl implements StoreDaoCustom {
         return targetStore;
     }
 
+    @Override
     public StoreId findStoreIdByUserEmail(String email) {
         QUser user = QUser.user;
         QStore store = QStore.store;
@@ -51,5 +59,8 @@ public class StoreDaoCustomImpl implements StoreDaoCustom {
         return storeId;
     }
 
-
+    @Override
+    public Optional<Store> findById(StoreId storeId) {
+        return storeJpaRepository.findById(storeId);
+    }
 }
