+++ api/src/main/java/com/albert/commerce/store/query/StoreDao.java
@@ -1,8 +1,14 @@
 package com.albert.commerce.store.query;
 
-import com.albert.commerce.store.infra.StoreDaoCustom;
-import com.albert.commerce.store.infra.StoreJpaRepository;
+import com.albert.commerce.store.command.domain.Store;
+import com.albert.commerce.store.command.domain.StoreId;
+import java.util.Optional;
 
-public interface StoreDao extends StoreJpaRepository, StoreDaoCustom {
+public interface StoreDao {
 
+    Optional<Store> findById(StoreId storeId);
+
+    Store findStoreByUserEmail(String email);
+
+    StoreId findStoreIdByUserEmail(String email);
 }
